Uncomment Uri encoding of parameters
Query string for initiating the access code retrieval was commented out. This results in inconsistent use of the parameters as they are plain values when getting the access code, but they are Uri encoded strings when the http client makes a post request to exchange the access code into an access token.